+++ src/test/java/net/snowflake/client/jdbc/ResultSetMultiTimeZoneLatestIT.java
@@ -3,7 +3,6 @@ package net.snowflake.client.jdbc;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-import java.sql.Connection;
 import java.sql.Date;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -19,8 +18,7 @@ import java.util.TimeZone;
 import net.snowflake.client.ConditionalIgnoreRule;
 import net.snowflake.client.RunningOnGithubAction;
 import net.snowflake.client.category.TestCategoryResultSet;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
@@ -55,21 +53,8 @@ public class ResultSetMultiTimeZoneLatestIT extends BaseJDBCTest {
     System.setProperty("user.timezone", timeZone);
   }
 
-  private static Connection connection;
-
-  @BeforeClass
-  public static void setUpConnection() throws SQLException {
-    connection = BaseJDBCTest.getConnection();
-  }
-
-  @AfterClass
-  public static void closeConnection() throws SQLException {
-    if (connection != null && !connection.isClosed()) {
-      connection.close();
-    }
-  }
-
-  public Connection init() throws SQLException {
+  @Before
+  public void init() throws SQLException {
     try (Statement statement = connection.createStatement()) {
       statement.execute(
           "alter session set "
@@ -81,7 +66,6 @@ public class ResultSetMultiTimeZoneLatestIT extends BaseJDBCTest {
               + "TIMESTAMP_NTZ_OUTPUT_FORMAT='DY, DD MON YYYY HH24:MI:SS TZHTZM'");
       statement.execute("alter session set jdbc_query_result_format = '" + queryResultFormat + "'");
     }
-    return connection;
   }
 
   /**
@@ -92,7 +76,6 @@ public class ResultSetMultiTimeZoneLatestIT extends BaseJDBCTest {
    */
   @Test
   public void testTimesWithGetTimestamp() throws SQLException {
-    Connection connection = init();
     try (Statement statement = connection.createStatement()) {
       String timeStringValue = "10:30:50.123456789";
       String timestampStringValue = "1970-01-01 " + timeStringValue;
@@ -127,7 +110,6 @@ public class ResultSetMultiTimeZoneLatestIT extends BaseJDBCTest {
    */
   @Test
   public void testTimestampNTZWithDaylightSavings() throws SQLException {
-    Connection connection = init();
     try (Statement statement = connection.createStatement()) {
       statement.execute(
           "alter session set TIMESTAMP_TYPE_MAPPING='TIMESTAMP_NTZ'," + "TIMEZONE='Europe/London'");
@@ -149,7 +131,6 @@ public class ResultSetMultiTimeZoneLatestIT extends BaseJDBCTest {
     Calendar cal = null;
     SimpleDateFormat sdf = null;
 
-    Connection connection = init();
     try (Statement statement = connection.createStatement()) {
       statement.execute("alter session set JDBC_FORMAT_DATE_WITH_TIMEZONE=true");
       try (ResultSet rs =
@@ -246,7 +227,6 @@ public class ResultSetMultiTimeZoneLatestIT extends BaseJDBCTest {
    * @throws SQLException
    */
   private void testUseSessionTimeZoneHelper(boolean useDefaultParamSettings) throws SQLException {
-    Connection connection = init();
     try (Statement statement = connection.createStatement()) {
       try {
         // create table with all timestamp types, time, and date
