+++ src/test/java/net/snowflake/client/jdbc/ResultSetVectorLatestIT.java
@@ -5,7 +5,6 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
-import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
@@ -14,8 +13,6 @@ import java.sql.Types;
 import java.util.Arrays;
 import java.util.List;
 import net.snowflake.client.category.TestCategoryResultSet;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
@@ -44,23 +41,9 @@ public class ResultSetVectorLatestIT extends ResultSet0IT {
     return Arrays.asList("json", "arrow");
   }
 
-  private static Connection con;
-
-  @BeforeClass
-  public static void setUpConnection() throws SQLException {
-    con = BaseJDBCTest.getConnection();
-  }
-
-  @AfterClass
-  public static void closeConnection() throws SQLException {
-    if (con != null && !con.isClosed()) {
-      con.close();
-    }
-  }
-
   @Test
   public void testGetIntVectorAsIntArray() throws SQLException {
-    try (Statement stmt = con.createStatement()) {
+    try (Statement stmt = connection.createStatement()) {
       enforceQueryResultFormat(stmt);
       Integer[] vector = {-1, 5};
       try (ResultSet resultSet = stmt.executeQuery("select " + vectorToString(vector, "int"))) {
@@ -75,7 +58,7 @@ public class ResultSetVectorLatestIT extends ResultSet0IT {
 
   @Test
   public void testGetIntVectorAsLongArray() throws SQLException {
-    try (Statement stmt = con.createStatement()) {
+    try (Statement stmt = connection.createStatement()) {
       enforceQueryResultFormat(stmt);
       Long[] vector = {-1L, 5L};
       try (ResultSet resultSet = stmt.executeQuery("select " + vectorToString(vector, "int"))) {
@@ -89,7 +72,7 @@ public class ResultSetVectorLatestIT extends ResultSet0IT {
 
   @Test
   public void testGetFloatVectorAsFloatArray() throws SQLException {
-    try (Statement stmt = con.createStatement()) {
+    try (Statement stmt = connection.createStatement()) {
       enforceQueryResultFormat(stmt);
       Float[] vector = {-1.2f, 5.1f, 15.87f};
       try (ResultSet resultSet = stmt.executeQuery("select " + vectorToString(vector, "float"))) {
@@ -103,7 +86,7 @@ public class ResultSetVectorLatestIT extends ResultSet0IT {
 
   @Test
   public void testGetNullAsIntVector() throws SQLException {
-    try (Statement stmt = con.createStatement()) {
+    try (Statement stmt = connection.createStatement()) {
       enforceQueryResultFormat(stmt);
       try (ResultSet resultSet = stmt.executeQuery("select null::vector(int, 2)")) {
         assertTrue(resultSet.next());
@@ -117,7 +100,7 @@ public class ResultSetVectorLatestIT extends ResultSet0IT {
 
   @Test
   public void testGetNullAsFloatVector() throws SQLException {
-    try (Statement stmt = con.createStatement()) {
+    try (Statement stmt = connection.createStatement()) {
       enforceQueryResultFormat(stmt);
       try (ResultSet resultSet = stmt.executeQuery("select null::vector(float, 2)")) {
         assertTrue(resultSet.next());
@@ -131,7 +114,7 @@ public class ResultSetVectorLatestIT extends ResultSet0IT {
 
   @Test
   public void testGetIntVectorFromTable() throws SQLException {
-    try (Statement stmt = con.createStatement()) {
+    try (Statement stmt = connection.createStatement()) {
       enforceQueryResultFormat(stmt);
       stmt.execute("create or replace table test_vector_int(x vector(int, 2), y int)");
       stmt.execute("insert into test_vector_int select [3, 7]::vector(int, 2), 15");
@@ -147,7 +130,7 @@ public class ResultSetVectorLatestIT extends ResultSet0IT {
 
   @Test
   public void testGetFloatVectorFromTable() throws SQLException {
-    try (Statement stmt = con.createStatement()) {
+    try (Statement stmt = connection.createStatement()) {
       enforceQueryResultFormat(stmt);
       stmt.execute("create or replace table test_vector_float(x vector(float, 2), y float)");
       stmt.execute("insert into test_vector_float select [-3, 7.1]::vector(float, 2), 20.3");
@@ -163,7 +146,7 @@ public class ResultSetVectorLatestIT extends ResultSet0IT {
   /** Added in > 3.16.1 */
   @Test
   public void testGetVectorViaGetStringIsEqualToTheGetObject() throws SQLException {
-    try (Statement stmt = con.createStatement()) {
+    try (Statement stmt = connection.createStatement()) {
       enforceQueryResultFormat(stmt);
       Integer[] intVector = {-1, 5};
       Float[] floatVector = {-1.2f, 5.1f, 15.87f};
