+++ src/test/java/net/snowflake/client/jdbc/StatementLatestIT.java
@@ -29,8 +29,6 @@ import net.snowflake.client.category.TestCategoryStatement;
 import net.snowflake.client.core.ParameterBindingDTO;
 import net.snowflake.client.core.SFSession;
 import net.snowflake.client.core.bind.BindUploader;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
@@ -46,20 +44,6 @@ import org.junit.rules.TemporaryFolder;
 public class StatementLatestIT extends BaseJDBCTest {
   protected static String queryResultFormat = "json";
 
-  private static Connection con;
-
-  @BeforeClass
-  public static void setUpConnection() throws SQLException {
-    con = getConnection();
-  }
-
-  @AfterClass
-  public static void closeConnection() throws SQLException {
-    if (con != null && !con.isClosed()) {
-      con.close();
-    }
-  }
-
   public static Connection getConnection() throws SQLException {
     Connection conn = BaseJDBCTest.getConnection();
     try (Statement stmt = conn.createStatement()) {
@@ -72,7 +56,7 @@ public class StatementLatestIT extends BaseJDBCTest {
 
   @Test
   public void testExecuteCreateAndDrop() throws SQLException {
-    try (Statement statement = con.createStatement()) {
+    try (Statement statement = connection.createStatement()) {
 
       boolean success = statement.execute("create or replace table test_create(colA integer)");
       assertFalse(success);
@@ -193,7 +177,7 @@ public class StatementLatestIT extends BaseJDBCTest {
    */
   @Test
   public void testExecuteOpenResultSets() throws SQLException {
-    try (Statement statement = con.createStatement()) {
+    try (Statement statement = connection.createStatement()) {
       for (int i = 0; i < 10; i++) {
         statement.execute("select 1");
         statement.getResultSet();
@@ -202,7 +186,7 @@ public class StatementLatestIT extends BaseJDBCTest {
       assertEquals(9, statement.unwrap(SnowflakeStatementV1.class).getOpenResultSets().size());
     }
 
-    try (Statement statement = con.createStatement()) {
+    try (Statement statement = connection.createStatement()) {
       for (int i = 0; i < 10; i++) {
         statement.execute("select 1");
         ResultSet resultSet = statement.getResultSet();
@@ -259,7 +243,7 @@ public class StatementLatestIT extends BaseJDBCTest {
     String schemaName =
         TestUtil.GENERATED_SCHEMA_PREFIX
             + SnowflakeUtil.randomAlphaNumeric(255 - TestUtil.GENERATED_SCHEMA_PREFIX.length());
-    try (Statement stmt = con.createStatement()) {
+    try (Statement stmt = connection.createStatement()) {
       stmt.execute("create schema " + schemaName);
       stmt.execute("use schema " + schemaName);
       stmt.execute("drop schema " + schemaName);
@@ -269,7 +253,7 @@ public class StatementLatestIT extends BaseJDBCTest {
   /** Added in > 3.14.4 */
   @Test
   public void testQueryIdIsSetOnFailedQueryExecute() throws SQLException {
-    try (Statement stmt = con.createStatement()) {
+    try (Statement stmt = connection.createStatement()) {
       assertNull(stmt.unwrap(SnowflakeStatement.class).getQueryID());
       try {
         stmt.execute("use database not_existing_database");
@@ -285,7 +269,7 @@ public class StatementLatestIT extends BaseJDBCTest {
   /** Added in > 3.14.4 */
   @Test
   public void testQueryIdIsSetOnFailedExecuteUpdate() throws SQLException {
-    try (Statement stmt = con.createStatement()) {
+    try (Statement stmt = connection.createStatement()) {
       assertNull(stmt.unwrap(SnowflakeStatement.class).getQueryID());
       try {
         stmt.executeUpdate("update not_existing_table set a = 1 where id = 42");
@@ -301,7 +285,7 @@ public class StatementLatestIT extends BaseJDBCTest {
   /** Added in > 3.14.4 */
   @Test
   public void testQueryIdIsSetOnFailedExecuteQuery() throws SQLException {
-    try (Statement stmt = con.createStatement()) {
+    try (Statement stmt = connection.createStatement()) {
       assertNull(stmt.unwrap(SnowflakeStatement.class).getQueryID());
       try {
         stmt.executeQuery("select * from not_existing_table");
